+++ src/main/java/net/spell_engine/client/gui/HudKeyVisuals.java
@@ -6,7 +6,7 @@ import net.spell_engine.SpellEngineMod;
 import java.util.Map;
 
 public class HudKeyVisuals {
-    private static final Drawable.Texture mouseTexture = new Drawable.Texture(new Identifier(SpellEngineMod.ID, "textures/gui/mouse.png"), 64, 64);
+    private static final Drawable.Texture mouseTexture = new Drawable.Texture(new Identifier(SpellEngineMod.ID, "textures/gui/hotkeys.png"), 256, 256);
     public static Map<String, Drawable.Component> custom = Map.ofEntries(
             Map.entry("key.mouse.left", new Drawable.Component(
                     new Drawable.Draw(0, 0, 10, 12),
@@ -28,5 +28,19 @@ public class HudKeyVisuals {
                     new Drawable.Draw(16, 16, 10, 12),
                     mouseTexture
             ))
+    );
+
+    private static final int keyboardV = 48;
+    public static final Drawable.Component buttonLeading = new Drawable.Component(
+            new Drawable.Draw(0, keyboardV, 3, 12),
+            mouseTexture
+    );
+    public static final Drawable.Component buttonCenter = new Drawable.Component(
+            new Drawable.Draw(16, keyboardV, 16, 12),
+            mouseTexture
+    );
+    public static final Drawable.Component buttonTrailing= new Drawable.Component(
+            new Drawable.Draw(16-3, keyboardV, 3, 12),
+            mouseTexture
     );
 }